fixed hotel

+++ Travel-Agency-Logic/Services/Hotel.cs
@@ -1,10 +1,12 @@
 using Microsoft.EntityFrameworkCore;
+using Travel_Agency_Core;
 using Travel_Agency_DataBase;
 using Travel_Agency_Domain.Services;
+using Travel_Agency_Logic.Core;
 
 namespace Travel_Agency_Logic.Services
 {
-    public class HotelService
+    public class HotelService : IHotelService
     {
         private readonly TravelAgencyContext _context;
 
@@ -13,33 +15,40 @@ namespace Travel_Agency_Logic.Services
             _context = context;
         }
 
-        public async Task<IEnumerable<Hotel>> GetAllHotelsAsync()
+        public async Task<ApiResponse<Hotel>> CreateHotel(Hotel hotel, UserBasic user)
         {
-            return await _context.Hotels.ToListAsync();
-        }
-
-        public async Task<Hotel?> GetHotelByIdAsync(int id)
-        {
-            return await _context.Hotels.FindAsync(id);
-        }
-
-        public async Task CreateHotelAsync(Hotel hotel)
-        {
+            if (!CheckPermissions(user))
+                return new Unauthorized<Hotel>("You don't have permissions");   
+            if (!await _context.Agencies.AnyAsync(a => a.Id == user.Id))
+                return new NotFound<Hotel>("Hotel not found");
             _context.Hotels.Add(hotel);
             await _context.SaveChangesAsync();
+            return new ApiResponse<Hotel>(hotel);
         }
 
-        public async Task UpdateHotelAsync(Hotel hotel)
+        public async Task<ApiResponse<Hotel>> UpdateHotel(Hotel hotel, UserBasic user)
         {
-            _context.Entry(hotel).State = EntityState.Modified;
+            if (!CheckPermissions(user))
+                return new Unauthorized<Hotel>("You don't have permissions");  
+            if (!await _context.Hotels.AnyAsync(h => h.Id == hotel.Id))
+                return new NotFound<Hotel>("Hotel not found");
+            _context.Update(hotel).State = EntityState.Modified;
             await _context.SaveChangesAsync();
+            return new ApiResponse<Hotel>(hotel);
         }
 
-        public async Task DeleteHotelAsync(int id)
+        public async Task<ApiResponse<Hotel>> DeleteHotel(int id, UserBasic user)
         {
+            if (!CheckPermissions(user))
+                return new Unauthorized<Hotel>("You don't have permissions");  
+            if (!await _context.Hotels.AnyAsync(h => h.Id == id))
+                return new NotFound<Hotel>("Hotel not found");
             var hotel = await _context.Hotels.FindAsync(id);
             _context.Hotels.Remove(hotel!);
             await _context.SaveChangesAsync();
+            return new ApiResponse<Hotel>(hotel!);
         }
+
+        private static bool CheckPermissions(UserBasic user) => user.Role == Roles.AdminApp || user.Role == Roles.EmployeeApp;
     }
 }
